Convert to ViewEvents -> RoomDetailViewModel part 2
