+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailViewEvents.kt
@@ -16,7 +16,9 @@
 
 package im.vector.riotx.features.home.room.detail
 
+import androidx.annotation.StringRes
 import im.vector.riotx.core.platform.VectorViewEvents
+import im.vector.riotx.features.command.Command
 
 /**
  * Transient events for RoomDetail
@@ -29,4 +31,15 @@ sealed class RoomDetailViewEvents : VectorViewEvents {
     data class ActionFailure(val action: RoomDetailAction, val throwable: Throwable) : RoomDetailViewEvents()
 
     data class ShowMessage(val message: String) : RoomDetailViewEvents()
+
+    abstract class SendMessageResult : RoomDetailViewEvents()
+
+    object MessageSent : SendMessageResult()
+    class SlashCommandError(val command: Command) : SendMessageResult()
+    class SlashCommandUnknown(val command: String) : SendMessageResult()
+    data class SlashCommandHandled(@StringRes val messageRes: Int? = null) : SendMessageResult()
+    object SlashCommandResultOk : SendMessageResult()
+    class SlashCommandResultError(val throwable: Throwable) : SendMessageResult()
+    // TODO Remove
+    object SlashCommandNotImplemented : SendMessageResult()
 }
