--- vector/src/main/java/im/vector/riotx/features/home/room/detail/SendMessageResult.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2019 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package im.vector.riotx.features.home.room.detail
-
-import androidx.annotation.StringRes
-import im.vector.riotx.features.command.Command
-
-sealed class SendMessageResult {
-    object MessageSent : SendMessageResult()
-    class SlashCommandError(val command: Command) : SendMessageResult()
-    class SlashCommandUnknown(val command: String) : SendMessageResult()
-    data class SlashCommandHandled(@StringRes val messageRes: Int? = null) : SendMessageResult()
-    object SlashCommandResultOk : SendMessageResult()
-    class SlashCommandResultError(val throwable: Throwable) : SendMessageResult()
-    // TODO Remove
-    object SlashCommandNotImplemented : SendMessageResult()
-}
